Update LED Shield Practice
Updated footprints for flush fitting resistors and reran traces. Build 2nd board from last years run, changed led resistors from 500 to 2k, 500 burned out leds with 12v.
Board works but the clock led is lit and turns off when it receives a signal, the transistor is acting as a logic NOT, unsure how to debug, need time on the breadboard.