Cleanup Comments. Add TODOs
Finished the v3.2 build and it works! Noticed a few issues, zeros are not displaying probably from leaving the RBO and RBI (pins 3,4) on the 4033's floating, but need to test. I also left some comments about how I think we can make this count from 1 to 16, instead of from 0 (not showing) to 15. If I can fix the zeros displaying then maybe we don't have to deal with the extra pulse to skip "00" on the counters.